Add polyline stuff

+++ static/build.sh
@@ -4,8 +4,13 @@ cwd="$( cd "${BASH_SOURCE[0]%/*}" && pwd )"
 
 pushd "$cwd/.."
 .cabal-sandbox/bin/fay --include --include=../.cabal-sandbox --package fay-text,fay-dom "$cwd/mapview.hs"
-echo "Using Google's Closure Compiler with 'advanced optimizations'"
-echo "This might be slow on some computers..."
-closure-compiler --externs /usr/share/closure-compiler/externs/contrib/maps/google_maps_api_v3_15.js --compilation_level ADVANCED_OPTIMIZATIONS "$cwd/mapview.js" > "$cwd/mapview.min.js" 2>/dev/null
-rm "$cwd/mapview.js"
+if [ "$1" == "dev" ]; then
+  echo "[dev] Not sending to Closure Compiler to minify."
+  mv "$cwd/mapview.js" "$cwd/mapview.min.js"
+else
+  echo "Using Google's Closure Compiler with 'advanced optimizations'"
+  echo "This might be slow on some computers..."
+  closure-compiler --externs /usr/share/closure-compiler/externs/contrib/maps/google_maps_api_v3_15.js --compilation_level ADVANCED_OPTIMIZATIONS "$cwd/mapview.js" > "$cwd/mapview.min.js" 2>/dev/null
+  rm "$cwd/mapview.js"
+fi
 popd
